+++ src/Crowdin.Api/Distributions/Distribution.cs
@@ -21,10 +21,10 @@ namespace Crowdin.Api.Distributions
         public string Name { get; set; }
 
         [JsonProperty("fileIds")]
-        public int[] FileIds { get; set; }
+        public long[] FileIds { get; set; }
 
         [JsonProperty("bundleIds")]
-        public int[] BundleIds { get; set; }
+        public long[] BundleIds { get; set; }
         
         [JsonProperty("createdAt")]
         public DateTimeOffset CreatedAt { get; set; }
